refactor: dirtributed view functions
