+++ patients/urls.py
@@ -1,10 +1,14 @@
 from django.urls import path, include
 
-from .views import (appointment_detail, appointment_list, dashboard,
-                    doctor_detail, doctor_list, patient_list, patient_detail,
-                    invoice_list, invoice_detail, invoice_print,
-                    get_monthly_sales_chart_data, stats,
-                    get_daily_sales_chart_data)
+from .views_file import (get_monthly_sales_chart_data,
+                         get_daily_sales_chart_data)
+
+from patients.views.dashboard import dashboard
+from patients.views.patient import patient_detail, patient_list
+from patients.views.doctor import doctor_detail, doctor_list
+from patients.views.appointment import appointment_detail, appointment_list
+from patients.views.invoice import invoice_detail, invoice_list, invoice_print
+from patients.views.stats import stats
 
 urlpatterns = [
     path('', dashboard, name='index'),
