+++ patients/views/__init__.py
@@ -1,0 +1,5 @@
+from django.core.cache.backends.base import DEFAULT_TIMEOUT
+from django.conf import settings
+
+CACHE_TTL = getattr(settings, 'CACHE_TTL', DEFAULT_TIMEOUT)
+LOGIN_URL = getattr(settings, 'LOGIN_URL')
